refactor(tests): replace createTestApp utility with createDiskApp and createMemoryApp for improved clarity and separation of concerns in project tests chore(tests): remove deprecated createTestApp utility and related code to streamline test setup and enhance maintainability
