+++ server/test/projects/projects-crud-disk.spec.ts
@@ -1,26 +1,23 @@
 import {ProjectDto} from '@/entities/project.entity';
-import request from 'supertest';
-import {createTestApp, CreateTestAppResult} from '../utils/create-test-app';
+import {CreateDisk, createDiskApp} from '../utils/create-disk-app';
 
 describe('Projects', () => {
-    let testApp: CreateTestAppResult;
-    let r: ReturnType<typeof request>;
+    let app: CreateDisk;
 
     beforeAll(async () => {
-        testApp = await createTestApp({mode: 'disk'});
-        r = request(testApp.app.getHttpServer());
+        app = await createDiskApp();
     });
 
     afterAll(async () => {
-        await testApp.shutdown();
+        await app.shutdown();
     });
 
     it('should have no projects', () => {
-        return r.get('/api/projects').expect(200).expect([]);
+        return app.request.get('/api/projects').expect(200).expect([]);
     });
 
     it('should not open a project when creating it', () => {
-        return r
+        return app.request
             .post('/api/projects')
             .send({name: 'test'})
             .expect(201)
